+++ R/get_esi.R
@@ -1,15 +1,17 @@
-#' Get CHIRPS precipitation data
+#' Get Evaporative Stress Index (ESI) data
 #' 
-#' Get daily precipitation data from the "Climate Hazards Group InfraRed Precipitation 
-#' with Station Data" via ClimateSERV API Client. ClimateSERV works with geojson 
-#' of type 'Polygon'. The input \code{object} is then transformed into polygons
-#' with a small buffer area around the point.
+#' Get Evaporative Stress Index (ESI) from SERVIR Global
+#' via ClimateSERV API Client. ESI is available every four (or twelve) weeks from 2001 to present.
+#' The dataset may contain cloudy data which is returned as NAs.
+#' ClimateSERV works with geojson of type 'Polygon'. The input \code{object} is then 
+#' transformed into polygons with a small buffer area around the point.
 #' 
 #' @param object input, an object of class \code{\link{data.frame}} or \code{\link{sf}}
 #' @param dates a character of start and end dates in that order in the format
 #'  YYYY-MM-DD
 #' @param operation optional, an integer that represents which type of
 #' statistical operation to perform on the dataset
+#' @param period an integer value for the period of ESI data, four weeks period = 1, twelve weeks = 2
 #' @param as.sf logical, returns an object of class "sf" for S3 method of class \code{\link{sf}}
 #' @param ... further arguments passed to \code{sf} methods. See details 
 #' @details  
@@ -19,101 +21,93 @@
 #' 
 #' nQuadSegs: integer, number of segments per buffer quadrant
 #' 
-#' @return A data frame of CHIRPS data:
+#' @return A data frame of ESI data:
 #' \item{id}{the index for the rows in \code{object}}
-#' \item{dates}{the dates from which CHIRPS was requested}
+#' \item{dates}{the dates from which ESI was requested}
 #' \item{lon}{the longitude as provided in \code{object}}
 #' \item{lat}{the latitude as provided in \code{object}}
-#' \item{chirps}{the CHIRPS value in mm}
+#' \item{esi}{the ESI value}
 #' @references 
-#' 
-#' Funk C. et al. (2015). Scientific Data, 2, 150066.
-#'  <https://doi.org/10.1038/sdata.2015.66>
 #' 
 #' ClimateSERV <https://climateserv.servirglobal.net>
 #' 
+#' Evaporative Stress Index <http://catalogue.servirglobal.net/Product?product_id=198>
+#' 
 #' @examples
 #' \donttest{
-#'  
-#' # Three points in the Tapajos National Forest, Brazil
 #' library("chirps")
 #' 
 #' lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
 #'                      lat = c(-2.8094, -2.8756, -3.5279))
 #' 
-#' dates <- c("2017-12-15","2017-12-31")
+#' dates <- c("2017-12-15","2018-06-20")
 #' 
+#' # by default the function set a very small buffer around the points
+#' # which can return NAs due to cloudiness in ESI data
 #' 
-#' get_chirps(lonlat, dates)
+#' dat <- get_esi(lonlat, dates = dates)
 #' 
+#' dat
 #' 
-#' # S3 method for objects of class 'sf'
-#' library("sf")
+#' # the argument dist passed through sf increase the buffer area
 #' 
-#' lonlat <- data.frame(lon = c(-55.0281,-54.9857, -55.0714),
-#'                      lat = c(-2.8094, -2.8756, -3.5279))
+#' dat <- get_esi(lonlat, dates = dates, dist = 0.1)
 #' 
-#' lonlat <- st_as_sf(lonlat, coords = c("lon","lat"))
-#' 
-#' dates <- c("2017-12-15","2017-12-31")
-#' 
-#' get_chirps(lonlat, dates)
-#' 
-#' # as.sf TRUE returns an object of class 'sf'
-#' get_chirps(lonlat, dates, as.sf = TRUE)
+#' dat
 #' } 
-#' @import sf
-#' @import methods
-#' @importFrom jsonlite fromJSON
-#' @importFrom tibble as_tibble
 #' @export
-get_chirps <- function(object, dates, operation = 5, ...) {
+get_esi <- function(object, dates, operation = 5, period = 1, ...) {
   
-  UseMethod("get_chirps")
+  UseMethod("get_esi")
   
 }
 
-#' @rdname get_chirps
+#' @rdname get_esi
 #' @export
-get_chirps.default <- function(object, dates, operation = 5, ...) {
+get_esi.default <- function(object, dates, operation = 5, period = 1, ...) {
   
   # validate lonlat to check if they are within the CHIRPS range lat -50, 50
   .validate_lonlat(object, xlim = c(-180, 180), ylim = c(-50, 50))
   
   # validate dates
-  dates_inter <- .reformat_dates(dates, availability = c("1981-01-01", "0"))
-
+  dates_inter <- .reformat_dates(dates, availability = c("2001-01-01", "0"))
+  
   # get geojson strings from data.frame
   gj <- .dataframe_to_geojson(object, ...)
+  
+  if (period == 1) {
+    datatype <- 29
+  }
+  if (period == 2) {
+    datatype <- 33
+  }
   
   result <- .GET(gjson = gj, 
                  dates = dates_inter, 
                  operation = operation, 
-                 datatype = 0)
+                 datatype = datatype)
   
-  names(result)[names(result) == "value"] <- "chirps"
+  names(result)[names(result) == "value"] <- "esi"
   
   object$id <- rownames(object)
   
-  result <- merge(result, object, by = "id")
+  result <- merge(result, object, by = "id", all.y = TRUE)
   
-  names(result)[3:5] <- c("chirps", "lon", "lat")
+  names(result)[3:5] <- c("esi", "lon", "lat")
   
-  result <- result[, c("id", "lon", "lat", "date", "chirps")]
+  result <- result[, c("id", "lon", "lat", "date", "esi")]
   
   result <- tibble::as_tibble(result)
-  
-  class(result) <- c("chirps", class(result))
   
   return(result)
   
 }
 
-#' @rdname get_chirps
-#' @method get_chirps sf
+#' @rdname get_esi
+#' @method get_esi sf
 #' @export
-get_chirps.sf <- function(object, dates, operation = 5, as.sf = FALSE, ...) {
-
+get_esi.sf <- function(object, dates, operation = 5, period = 1, as.sf = FALSE, ...) {
+  
   # convert sf into a data.frame
   n <- nrow(object)
   
@@ -136,10 +130,17 @@ get_chirps.sf <- function(object, dates, operation = 5, as.sf = FALSE, ...) {
   # get geojson strings from data.frame
   gj <- .sf_to_geojson(object, ...)
   
+  if (period == 1) {
+    datatype <- 29
+  }
+  if (period == 2) {
+    datatype <- 33
+  }
+  
   result <- .GET(gjson = gj, 
                  dates = dates_inter, 
                  operation = operation, 
-                 datatype = 0)
+                 datatype = datatype)
   
   if (as.sf) {
     
@@ -163,13 +164,12 @@ get_chirps.sf <- function(object, dates, operation = 5, as.sf = FALSE, ...) {
     
     result <- merge(result, lonlat, by = "id")
     
-    names(result)[3:5] <- c("chirps", "lon", "lat")
+    names(result)[3:5] <- c("esi", "lon", "lat")
     
-    result <- result[, c("id", "lon", "lat", "date", "chirps")]
+    result <- result[, c("id", "lon", "lat", "date", "esi")]
     
     result <- tibble::as_tibble(result)
-    
-    class(result) <- c("chirps", class(result))
+
     
   }